+++ src/LaravelHexagonalServiceProvider.php
@@ -21,6 +21,12 @@ class LaravelHexagonalServiceProvider extends ServiceProvider
             __DIR__ . '/../config/hexagonal.php' => config_path('hexagonal.php'),
         ], 'hexagonal-config');
 
+        $this->publishes([
+            __DIR__ . '/../stubs/module' => base_path('stubs/hexagonal/module'),
+            __DIR__ . '/../stubs/shared' => base_path('stubs/hexagonal/shared'),
+            __DIR__ . '/../stubs/deptrac.yaml' => base_path('stubs/hexagonal/deptrac.yaml'),
+        ], 'hexagonal-stubs');
+
         if ($this->app->runningInConsole()) {
             $this->commands([
                 InstallCommand::class,